requirements.txt
neo4j
pydantic
pyyaml
pandas
numpy
openpyxl
xlrd<2.0
